requirements.txt
flask==2.0.1
pymongo==4.1.1
requests==2.27.1
beautifulsoup4==4.10.0
python-dotenv==0.19.1
werkzeug==2.0.3
